cli: add `--mount` to `docker run`
