+++ contrib/completion/bash/docker
@@ -1268,6 +1268,7 @@ _docker_container_run() {
 		--memory-swap
 		--memory-swappiness
 		--memory-reservation
+		--mount
 		--name
 		--network
 		--network-alias
